Frame-based -> moment-based
Up until now, the time between each rendered moment was the time
it took to reach the next visual frame.
That is, a virtual moment and a visual frame were being
conflated.
From a scientific perspective, this is an unreliable
discretization.
Each moment should be separated from the previous one by a
uniform amount of virtual time.